Server_side/requirements.txt
absl-py==1.3.0
asgiref==3.6.0
astroid==2.15.0
astunparse==1.6.3
autopep8==1.6.0
Babel==2.12.1
cachetools==5.2.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
channels==4.0.0
charset-normalizer==3.1.0
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.28
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
docstring-to-markdown==0.11
docx==0.2.4
docxcompose==1.4.0
docxtpl==0.16.6
easydict==1.10
et-xmlfile==1.1.0
flake8==6.0.0
Flask==2.2.2
flatbuffers==23.3.3
fonttools==4.38.0
gast==0.4.0
gensim==4.2.0
google-auth==2.15.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
hiddenlayer==0.3
idna==3.4
importlib-metadata==5.1.0
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.8
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mccabe==0.7.0
ml-dtypes==0.1.0
mplfinance==0.12.9b7
mysqlclient==2.1.1
netron==6.2.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.0.0
prettytable==3.5.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydocstyle==6.2.3
pyecharts==1.9.1
pyflakes==3.0.1
pygame==2.1.2
PyJWT @ file:///C:/ci/pyjwt_1657511236979/work
pylint==2.17.0
PyMySQL==1.0.2
pyparsing==3.0.9
PyQRCode==1.2.1
python-dateutil==2.8.2
python-docx==0.8.11
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
pytoolconfig==1.2.5
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rope==1.7.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
simplejson==3.18.0
six==1.16.0
smart-open==6.2.0
snowballstemmer==2.2.0
snownlp==0.12.3
sqlparse==0.4.3
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tomlkit==0.11.6
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.5.0
tzdata==2023.2
urllib3==1.26.15
Werkzeug==2.2.2
whatthepatch==1.0.4
wincertstore==0.2
wrapt==1.14.1
yapf==0.32.0
zipp==3.11.0
